refactor: begin ripping apart simapp and simd binary wiring
