+++ middleware/packet-forward-middleware/testing/simapp/simd/cmd/cmd_test.go
@@ -1,0 +1,25 @@
+package cmd_test
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
+	"github.com/cosmos/cosmos-sdk/x/genutil/client/cli"
+
+	"github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v9/testing/simapp"
+	"github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v9/testing/simapp/simd/cmd"
+)
+
+func TestInitCmd(t *testing.T) {
+	rootCmd := cmd.NewRootCmd()
+	rootCmd.SetArgs([]string{
+		"init",        // Test the init cmd
+		"simapp-test", // Moniker
+		fmt.Sprintf("--%s=%s", cli.FlagOverwrite, "true"), // Overwrite genesis.json, in case it already exists
+	})
+
+	require.NoError(t, svrcmd.Execute(rootCmd, "", simapp.DefaultNodeHome))
+}
