+++ middleware/packet-forward-middleware/testing/simapp/simd/main.go
@@ -3,16 +3,16 @@ package main
 import (
 	"os"
 
-	"github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v9/testing/simapp"
-
 	"cosmossdk.io/log"
 
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
+
+	"github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v9/testing/simapp"
+	"github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v9/testing/simapp/simd/cmd"
 )
 
 func main() {
-	rootCmd := NewRootCmd()
-
+	rootCmd := cmd.NewRootCmd()
 	if err := svrcmd.Execute(rootCmd, "", simapp.DefaultNodeHome); err != nil {
 		log.NewLogger(rootCmd.OutOrStderr()).Error("failure when running app", "err", err)
 		os.Exit(1)
